requirements_engine.txt
pydantic>=2.0.0
google-generativeai
rich
pytest
python-dotenv
